Move wizard MVC into dedicated module
