+++ src/aiidalab_qe/app/wizard/__init__.py
@@ -1,0 +1,7 @@
+from .model import WizardModel
+from .wizard import Wizard
+
+__all__ = [
+    "Wizard",
+    "WizardModel",
+]
