+++ src/aiidalab_qe/app/wizard/wizard.py
@@ -8,7 +8,7 @@ from aiidalab_qe.app.structure import StructureStep, StructureStepModel
 from aiidalab_qe.app.submission import SubmissionStep, SubmissionStepModel
 from aiidalab_qe.common.wizard import State, WizardStep
 
-from .wizard_model import WizardModel
+from .model import WizardModel
 
 
 class Wizard(ipw.Accordion):
